+++ packages/ai-runtime/src/front/index.tsx
@@ -1,8 +1,12 @@
 export { getDefaultOutputComponent } from './constants';
 
-export { runAgent } from './api/session';
+export * from './api/agent';
+export * from './api/session';
+export * from './api/message';
 
 export * from './components/AgentErrorBoundary';
+export { default as ScrollView } from './components/ScrollView';
+
 export * from './contexts/ComponentPreferences';
 export * from './contexts/Agent';
 export * from './contexts/Api';